Always update network interfaces information in cache
`anet.Interfaces` and `net.Interfaces` use the same `zoneCache`
underneath (and the same `lastFetched`). But because `net.Interfaces`
is currently not fully functional on Android, it always tries to clear
the cache and update the `lastFetched` timestamp. In certain
situations, this may prevent `anet.Interfaces` from populating the
cache at all.

Example pseudocode:
  for {
    _, _ = net.Listen(...) // clears zoneCache, updates the timestamp
    _, _ = anet.Interfaces() // doesn't update the cache because it's up to date
  }

The `zoneCache` is used to reduce the cost of resolving IPv6
addressing scope zones. Because all expensive operations have already
been performed in `anet.Interfaces`, there is no reason not to update
the cache (even if it has recent timestamp).